requirements.txt
langgraph
langchain
openai
groq
faiss-cpu
redis
streamlit
python-dotenv 
sentence-transformers
numpy 